requirements.txt
datasets==3.6.0
langchain==0.3.26
langchain_core==0.3.67
langchain_openai==0.3.27
langchain-community==0.3.27
chromadb==1.5.1
nltk==3.9.1
sentence_transformers==5.0.0
streamlit==1.46.1
torch==2.7.1
faiss-cpu==1.13.2